+++ src/utils.rs
@@ -7,7 +7,7 @@ pub fn convert_mouse_position_to_world_coordinates(
     window: &Window,
     transform: &Transform,
     scaling_factor: f32,
-    floor: i32,
+    floor: i64,
     mouse_position: Vec2,
 ) -> Position {
     Vec3::new(
